+++ include/avtp/acf/FlexRay.h
@@ -113,7 +113,7 @@ static const Avtp_FieldDescriptor_t Avtp_FlexRayFieldDesc[AVTP_FLEXRAY_FIELD_MAX
  * @returns Value of the ACF message type field.
  */
 static inline uint8_t Avtp_FlexRay_GetAcfMsgType(const Avtp_FlexRay_t* const pdu) {
-    return GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_ACF_MSG_TYPE);
+    return (uint8_t) GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_ACF_MSG_TYPE);
 }
 
 /** 
@@ -123,7 +123,7 @@ static inline uint8_t Avtp_FlexRay_GetAcfMsgType(const Avtp_FlexRay_t* const pdu
  * @returns Value of the ACF message length field.
  */
 static inline uint16_t Avtp_FlexRay_GetAcfMsgLength(const Avtp_FlexRay_t* const pdu) {
-    return GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_ACF_MSG_LENGTH);
+    return (uint16_t) GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_ACF_MSG_LENGTH);
 }
 
 /**
@@ -133,7 +133,7 @@ static inline uint16_t Avtp_FlexRay_GetAcfMsgLength(const Avtp_FlexRay_t* const 
  * @returns Value of the ACF FlexRay PDU padding field.
  */
 static inline uint8_t Avtp_FlexRay_GetPad(const Avtp_FlexRay_t* const pdu) {
-    return GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_PAD);
+    return (uint8_t) GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_PAD);
 }
 
 /**
@@ -143,7 +143,7 @@ static inline uint8_t Avtp_FlexRay_GetPad(const Avtp_FlexRay_t* const pdu) {
  * @returns Value of the ACF FlexRay PDU MTV field.
  */
 static inline uint8_t Avtp_FlexRay_GetMtv(const Avtp_FlexRay_t* const pdu) {
-    return GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_MTV);
+    return (uint8_t) GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_MTV);
 }
 
 /**
@@ -153,7 +153,7 @@ static inline uint8_t Avtp_FlexRay_GetMtv(const Avtp_FlexRay_t* const pdu) {
  * @returns Value of the ACF FlexRay PDU FR Bus ID field.
  */
 static inline uint8_t Avtp_FlexRay_GetFrBusId(const Avtp_FlexRay_t* const pdu) {
-    return GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_FR_BUS_ID);
+    return (uint8_t) GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_FR_BUS_ID);
 }
 
 /**
@@ -163,7 +163,7 @@ static inline uint8_t Avtp_FlexRay_GetFrBusId(const Avtp_FlexRay_t* const pdu) {
  * @returns Value of the ACF FlexRay PDU Channel field.
  */
 static inline uint8_t Avtp_FlexRay_GetChan(const Avtp_FlexRay_t* const pdu) {
-    return GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_CHAN);
+    return (uint8_t) GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_CHAN);
 }
 
 /**
@@ -173,7 +173,7 @@ static inline uint8_t Avtp_FlexRay_GetChan(const Avtp_FlexRay_t* const pdu) {
  * @returns Value of the ACF FlexRay PDU STR field.
  */
 static inline uint8_t Avtp_FlexRay_GetStr(const Avtp_FlexRay_t* const pdu) {
-    return GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_STR);
+    return (uint8_t) GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_STR);
 }
 
 /**
@@ -183,7 +183,7 @@ static inline uint8_t Avtp_FlexRay_GetStr(const Avtp_FlexRay_t* const pdu) {
  * @returns Value of the ACF FlexRay PDU SYN field.
  */
 static inline uint8_t Avtp_FlexRay_GetSyn(const Avtp_FlexRay_t* const pdu) {
-    return GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_SYN);
+    return (uint8_t) GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_SYN);
 }
 
 /**
@@ -193,7 +193,7 @@ static inline uint8_t Avtp_FlexRay_GetSyn(const Avtp_FlexRay_t* const pdu) {
  * @returns Value of the ACF FlexRay PDU PRE field.
  */
 static inline uint8_t Avtp_FlexRay_GetPre(const Avtp_FlexRay_t* const pdu) {
-    return GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_PRE);
+    return (uint8_t) GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_PRE);
 }
 
 /**
@@ -203,7 +203,7 @@ static inline uint8_t Avtp_FlexRay_GetPre(const Avtp_FlexRay_t* const pdu) {
  * @returns Value of the ACF FlexRay PDU NFI field.
  */
 static inline uint8_t Avtp_FlexRay_GetNfi(const Avtp_FlexRay_t* const pdu) {
-    return GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_NFI);
+    return (uint8_t) GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_NFI);
 }
 
 /**
@@ -223,7 +223,7 @@ static inline uint64_t Avtp_FlexRay_GetMessageTimestamp(const Avtp_FlexRay_t* co
  * @returns Value of the ACF FlexRay PDU FR Frame ID field.
  */
 static inline uint16_t Avtp_FlexRay_GetFrFrameId(const Avtp_FlexRay_t* const pdu) {
-    return GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_FR_FRAME_ID);
+    return (uint16_t) GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_FR_FRAME_ID);
 }
 
 /**
@@ -233,7 +233,7 @@ static inline uint16_t Avtp_FlexRay_GetFrFrameId(const Avtp_FlexRay_t* const pdu
  * @returns Value of the ACF FlexRay PDU Cycle field.
  */
 static inline uint8_t Avtp_FlexRay_GetCycle(const Avtp_FlexRay_t* const pdu) {
-    return GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_CYCLE);
+    return (uint8_t) GET_FLEXRAY_FIELD(AVTP_FLEXRAY_FIELD_CYCLE);
 }
 
 /**
